+++ packages/transaction-controller/src/utils/batch.ts
@@ -284,6 +284,7 @@ async function addTransactionBatchWith7702(
   } = request;
 
   const {
+    atomic,
     batchId: batchIdOverride,
     disableUpgrade,
     from,
@@ -349,7 +350,13 @@ async function addTransactionBatchWith7702(
     ),
   );
 
-  const batchParams = generateEIP7702BatchTransaction(from, nestedTransactions);
+  const batchParams = generateEIP7702BatchTransaction(
+    from,
+    nestedTransactions,
+    {
+      atomic,
+    },
+  );
 
   const txParams: TransactionParams = {
     ...batchParams,